+++ pyincore/analyses/meandamage/meandamage.py
@@ -13,6 +13,9 @@ import collections
 
 class MeanDamage(BaseAnalysis):
     """
+    Computes the mean structural damage from discrete probabilities of damage states and mean damage factors in the
+    damage ratios table.
+
     Args:
         incore_client (IncoreClient): Service authentication.
     """
@@ -29,24 +32,27 @@ class MeanDamage(BaseAnalysis):
         """
         return {
             "name": "mean-damage",
-            "description": "calculate the mean and expected damage using damage ratio table",
+            "description": "The process for computing the structural damage uses mean damage and standard deviation values from damage"
+            "ratios tables. The four damage state probabilities are multiplied by the mean damage and aggregated to get the"
+            "Mean damage for each individual structure (building, bridge, waterfacility etc.).",
             "input_parameters": [
                 {
                     "id": "result_name",
                     "required": True,
-                    "description": "result dataset name",
+                    "description": "Base name of result dataset.",
                     "type": str,
                 },
                 {
                     "id": "damage_interval_keys",
                     "required": True,
-                    "description": "Column name of the damage interval must be four and ranged in order",
+                    "description": "Damage interval column that must contain exactly four values, "
+                    "arranged in order.",
                     "type": List[str],
                 },
                 {
                     "id": "num_cpu",
                     "required": False,
-                    "description": "If using parallel execution, the number of cpus to request",
+                    "description": "Number of cpus to request, when using parallel execution.",
                     "type": int,
                 },
             ],
@@ -54,7 +60,7 @@ class MeanDamage(BaseAnalysis):
                 {
                     "id": "damage",
                     "required": True,
-                    "description": "damage result that has damage intervals in it",
+                    "description": "Damage result dataset that has damage intervals in it.",
                     "type": [
                         "ergo:buildingDamageVer4",
                         "ergo:buildingDamageVer5",
@@ -74,7 +80,7 @@ class MeanDamage(BaseAnalysis):
                 {
                     "id": "dmg_ratios",
                     "required": True,
-                    "description": "Damage Ratios table",
+                    "description": "Damage Ratios table.",
                     "type": [
                         "ergo:buildingDamageRatios",
                         "ergo:bridgeDamageRatios",
@@ -88,7 +94,7 @@ class MeanDamage(BaseAnalysis):
             "output_datasets": [
                 {
                     "id": "result",
-                    "description": "CSV file of mean damage",
+                    "description": "CSV file of mean damage.",
                     "type": "ergo:meanDamage",
                 }
             ],
@@ -157,11 +163,11 @@ class MeanDamage(BaseAnalysis):
         return output
 
     def mean_damage_bulk_input(self, damage, dmg_ratio_tbl):
-        """Run analysis for mean damage calculation
+        """Run analysis for mean damage calculation.
 
         Args:
-            damage (obj): output of building/bridge/waterfacility/epn damage that has damage interval
-            dmg_ratio_tbl (list): damage ratio table
+            damage (obj): output of building/bridge/waterfacility/epn damage that has damage interval.
+            dmg_ratio_tbl (list): damage ratio table.
 
         Returns:
             list: A list of ordered dictionaries with mean damage, deviation, and other data/metadata.
@@ -185,14 +191,14 @@ class MeanDamage(BaseAnalysis):
         return result
 
     def mean_damage(self, dmg, dmg_ratio_tbl, damage_interval_keys, is_bridge):
-        """Calculates mean damage based on damage probabilities and damage ratios
+        """Calculates mean damage based on damage probabilities and damage ratios.
 
         Args:
-            dmg (obj): dmg analysis output for a single entity in the built environment
+            dmg (obj): dmg analysis output for a single entity in the built environment.
             dmg_ratio_tbl (list): dmg ratio table.
-            damage_interval_keys (list): damage interval keys
+            damage_interval_keys (list): damage interval keys.
             is_bridge (bool): a boolean to indicate if the inventory type is bridge.
-            Bridge has its own way of calculating mean damage
+            Bridge has its own way of calculating mean damage.
 
         Returns:
             OrderedDict: A dictionary with mean damage, deviation, and other data/metadata.