+++ pyincore/analyses/mlenabledcgeslc/mlcgeslc.py
@@ -17,6 +17,17 @@ logger = pyglobals.LOGGER
 
 
 class MlEnabledCgeSlc(CoreCGEML):
+    """
+    The "Machine Learning Enabled Computable General Equilibrium (CGE) - Joplin" analysis merges advanced machine
+    learning with traditional CGE models to offer unprecedented insights into the economic impacts of disaster
+    scenarios on Joplin. Trained on a comprehensive dataset of numerous simulated disasters and their economic
+    effects, this hybrid approach excels in predicting the intricate dynamics of the city's economy under
+    various crises.
+
+    Args:
+        CoreCGEML (class): The base class for the ML enabled CGE model.
+    """
+
     model = "Machine Learning Enabled Computable General Equilibrium - Salt Lake City "
 
     # Coefficients files
@@ -98,7 +109,7 @@ class MlEnabledCgeSlc(CoreCGEML):
         )  # 4 labor groups
 
     def run(self) -> bool:
-        """Executes the ML enabled CGE model for Salt Lake City"""
+        """Executes the ML enabled CGE model for Salt Lake City."""
 
         logger.info(f"Running {self.model} model...")
         sector_shocks = pd.read_csv(
@@ -111,7 +122,7 @@ class MlEnabledCgeSlc(CoreCGEML):
             if sector.upper() not in [v.upper() for v in sector_shocks["sector"]]:
                 raise ValueError(
                     f"Sector {sector} not found in the sector shocks file with\n {sector_shocks['sector']} sectors.\n"
-                    + "Please make sure you have used the correct capital shocks"
+                    + "Please make sure you have used the correct capital shocks."
                 )
             shocks.append(
                 sector_shocks.loc[sector_shocks["sector"] == sector.upper()]["shock"]
@@ -132,12 +143,16 @@ class MlEnabledCgeSlc(CoreCGEML):
     def get_spec(self):
         return {
             "name": "Salt-Lake-cge",
-            "description": "CGE model for Salt Lake City.",
+            "description": "The Machine Learning Enabled Computable General Equilibrium (CGE) - Salt Lake City analysis merges "
+            "advanced machine learning with traditional CGE models to offer unprecedented insights into the economic impacts of "
+            "disaster scenarios on Salt Lake City. Trained on a comprehensive dataset of numerous simulated disasters and their "
+            "economic effects, this hybrid approach excels in predicting the intricate dynamics of the city’s economy "
+            "under various crises.",
             "input_parameters": [
                 {
                     "id": "result_name",
                     "required": False,
-                    "description": "Result CSV dataset name prefix",
+                    "description": "Result CSV dataset name prefix.",
                     "type": str,
                 }
             ],
@@ -145,7 +160,7 @@ class MlEnabledCgeSlc(CoreCGEML):
                 {
                     "id": "sector_shocks",
                     "required": True,
-                    "description": "Aggregation of building functionality states to capital shocks per sector",
+                    "description": "Aggregation of building functionality states to capital shocks per sector.",
                     "type": ["incore:capitalShocks"],
                 }
             ],
@@ -153,31 +168,31 @@ class MlEnabledCgeSlc(CoreCGEML):
                 {
                     "id": "domestic-supply",
                     "parent_type": "",
-                    "description": "CSV file of resulting domestic supply",
+                    "description": "CSV file of resulting domestic supply.",
                     "type": "incore:Employment",
                 },
                 {
                     "id": "gross-income",
                     "parent_type": "",
-                    "description": "CSV file of resulting gross income",
+                    "description": "CSV file of resulting gross income.",
                     "type": "incore:Employment",
                 },
                 {
                     "id": "pre-disaster-factor-demand",
                     "parent_type": "",
-                    "description": "CSV file of factor demand before disaster",
+                    "description": "CSV file of factor demand before disaster.",
                     "type": "incore:FactorDemand",
                 },
                 {
                     "id": "post-disaster-factor-demand",
                     "parent_type": "",
-                    "description": "CSV file of resulting factor-demand",
+                    "description": "CSV file of resulting factor-demand.",
                     "type": "incore:FactorDemand",
                 },
                 {
                     "id": "household-count",
                     "parent_type": "",
-                    "description": "CSV file of household count",
+                    "description": "CSV file of household count.",
                     "type": "incore:HouseholdCount",
                 },
             ],